#572 moved into their own module: - semaphore prototype - blocking queue - event queue - configuration settings
create new base object used to extend off of. relation and store uses this now.
