+++ src/utils/events.js
@@ -1,0 +1,3 @@
+import Backbone from 'backbone';
+
+export default Backbone.Events;
