+++ src/utils/extend.js
@@ -1,0 +1,3 @@
+import {Model} from 'backbone';
+
+export default Model.extend;
